Fixes the iteration counts in the DFTPM f-ADAPT data
This data is part of Figure S4 and the iteration count was not computed
correctly. Instead of reporting the number of f-ADAPT iteration (which
was limited to 4), the values prior to this commit were reporting the
number of Pauli terms in the adaptive ansatz (in q-ADAPT every iteration
adds a single such Pauli term, but in f-ADAPT this depends on the
excitation; e.g. a single excitation adds 4 Pauli terms, a double
excitation 8 Pauli terms).
This change does not affect any statements made in the paper or SI but
Figure S4 would look slightly different as the number of iterations
would be limited by (just like stated in the text).